starting new bounds structure
