+++ src/Coluna.jl
@@ -22,6 +22,8 @@ import Base: isempty, hash, isequal, length, iterate, getindex, lastindex,
     getkey, delete!, setindex!, haskey, copy, promote_rule, convert, isinteger,
     push!, filter, diff
 
+include("interface.jl")
+
 include("Containers/Containers.jl")
 using .Containers
