+++ src/Containers/Containers.jl
@@ -2,7 +2,9 @@ module Containers
 
 import ..Coluna
 
+import DataStructures
 import Primes
+import Printf
 import Base: <=, setindex!, get, getindex, haskey, keys, values, iterate, 
              length, lastindex, filter, show, keys, copy
 
@@ -13,6 +15,9 @@ export NestedEnum, @nestedenum, @exported_nestedenum,
 export getelements, getelement, rows, cols, columns, getrecords
 
 include("nestedenum.jl")
+include("solsandbounds.jl")
+
+# Following files will be deleted
 include("elements.jl")
 include("members.jl")
