+++ src/MathProg/MathProg.jl
@@ -72,7 +72,6 @@ include("variable.jl")
 include("constraint.jl")
 include("varconstr.jl")
 
-include("solsandbounds.jl")
 include("manager.jl")
 include("filters.jl")
 include("optimizationresults.jl")
